Explicit size for ports

+++ src/TelloRemoteControl.ino
@@ -1,6 +1,6 @@
 #include <WiFi.h>
 #include <WiFiUdp.h>
-
+#include <cstdint>
 
 // Debug Messages
 const String MESSAGE__FIRMWARE_INFO = "Tello Remote-Controller | v0.1";
@@ -13,8 +13,8 @@ const String STATUS_MESSAGE_RESPONSE__FAIL = "FAIL!";
 const String TELLO_WIFI_SSID = "TELLO-9F5E9A";
 const String TELLO_WIFI_PASSWORD = "";
 const String TELLO_IP = "192.168.10.1";
-const int TELLO_PORT = 8889;
-const int LOCAL_PORT = 9000;    // Local port to bind the UDP socket to. Can be any free port.
+const uint16_t TELLO_PORT = 8889;
+const uint16_t LOCAL_PORT = 9000;    // Local port to bind the UDP socket to. Can be any free port.
 
 bool initialized = false;
 byte buffer[1000];
